\name{internal.utils}


\alias{convert.counting}
\alias{convert.standard.counting}
\alias{hazard.simulation}
\alias{rhf.xvar.wt}
\alias{rhf.pooled.hazard}

\title{Internal Utility Functions}

\description{
  These are internal utility functions exported for advanced usage. This
  Rd file is used solely to register aliases.
}

\keyword{internal}
